Trying to make NSWindow independent (WIP)
It currently doesn't work. The window doesn't print if we don't
call poll_event. When we do it does show, but it segfaults not
long after...

We still have to investigate what is missing. Maybe we need
a delegate window controller but it would be nice to be able
to avoid that (and we don't know if that would solve anything).